move dataset functions into api
