+++ materials_commons/cli/dataset.py
@@ -7,109 +7,6 @@ from ..api import __api as mc_raw_api
 from materials_commons.cli.list_objects import ListObjects
 import materials_commons.cli.functions as clifuncs
 
-def get_all_datasets_from_experiment(project_id, experiment_id, remote=None):
-    remote = mc_raw_api.configure_remote(remote, None)
-    api_url = "projects/" + project_id + "/experiments/" + experiment_id + "/datasets"
-    result = mc_raw_api.get(remote.make_url_v2(api_url), remote)
-    for r in result:
-        r['name'] = r['title']
-    # return result
-    return [mcapi.Dataset(d) for d in result]
-
-def get_all_datasets_from_project(project_id, remote=None):
-    result = clifuncs.post_v3("listDatasets", {'project_id':project_id}, remote=remote)['data']
-    # print(json.dumps(result, indent=2))
-    for r in result:
-        r['name'] = r['title']
-    # return result
-    return [mcapi.Dataset(d) for d in result]
-
-def get_all_datasets_from_remote(remote=None):
-    result = clifuncs.post_v3("getTopViewedPublishedDatasets", remote=remote)['data']
-    for r in result:
-        r['name'] = r['title']
-    # return result
-    return [mcapi.Dataset(d) for d in result]
-
-def download_dataset_zipfile(dataset_id, output_file_path, remote=None):
-    remote = mc_raw_api.configure_remote(remote, None)
-    api_url = remote.make_url_v3("downloadDatasetZipfile")
-    params = {'apikey':remote.config.mcapikey, 'dataset_id':dataset_id}
-
-    with open(output_file_path, 'wb') as f:
-        r = requests.get(api_url, params=params, verify=False, stream=True)
-
-        if not r.ok:
-            r.raise_for_status()
-
-        for block in r.iter_content(8192):
-            if block:
-                f.write(block)
-
-    return output_file_path
-
-def create_dataset(project_id, title, description, remote=None):
-    result = clifuncs.post_v3(
-        "createDataset",
-        {
-            'project_id': project_id,
-            'title': title,
-            'description': description
-        },
-        remote=remote)['data']
-    return mcapi.Dataset(result)
-
-def delete_dataset(project_id, dataset_id, remote=None):
-    result = clifuncs.post_v3(
-        "deleteDataset",
-        {
-            'project_id': project_id,
-            'dataset_id': dataset_id
-        },
-        remote=remote)['data']
-    return result['success']
-
-def unpublish_dataset(project_id, dataset_id, remote=None):
-    """Unpublish a dataset
-
-    Returns
-    -------
-    dataset: mcapi.Dataset
-    """
-    result = clifuncs.post_v3(
-        "unpublishDataset",
-        {
-            'project_id': project_id,
-            'dataset_id': dataset_id
-        },
-        remote=remote)
-    return mcapi.Dataset(result['data'])
-
-def publish_dataset(project_id, dataset_id, remote=None):
-    """Publish a dataset
-
-    Returns
-    -------
-    dataset: mcapi.Dataset
-    """
-    result = clifuncs.post_v3(
-        "publishDataset",
-        {
-            'project_id': project_id,
-            'dataset_id': dataset_id
-        },
-        remote=remote)
-    return mcapi.Dataset(result['data'])
-
-def publish_private_dataset(project_id, dataset_id, remote=None):
-    result = clifuncs.post_v3(
-        "publishPrivateDataset",
-        {
-            'project_id': project_id,
-            'dataset_id': dataset_id
-        },
-        remote=remote)
-    return mcapi.Dataset(result['data'])
 
 class DatasetSubcommand(ListObjects):
 
@@ -125,18 +22,20 @@ class DatasetSubcommand(ListObjects):
             creatable=True,
             custom_selection_actions=['down', 'unpublish', 'publish', 'publish_private'],
             request_confirmation_actions={
-                'publish': 'Are you sure you want to publicly publish these datasets?'
+                'publish': 'Are you sure you want to publicly publish these datasets?',
+                'unpublish': 'Are you sure you want to unpublish these datasets?',
+                'publish_private': 'Are you sure you want to privately publish these datasets?'
             }
         )
 
     def get_all_from_experiment(self, expt):
-        return get_all_datasets_from_experiment(expt.project.id, expt.id, expt.project.remote)
+        return mcapi.get_all_datasets_from_experiment(expt.project.id, expt.id, expt.project.remote)
 
     def get_all_from_project(self, proj):
-        return get_all_datasets_from_project(proj.id, proj.remote)
+        return mcapi.get_all_datasets_from_project(proj.id, proj.remote)
 
     def get_all_from_remote(self, remote):
-        return get_all_datasets_from_remote(remote)
+        return mcapi.get_all_datasets_from_remote(remote)
 
     def list_data(self, obj):
 
@@ -177,11 +76,15 @@ class DatasetSubcommand(ListObjects):
 
     def add_custom_options(self, parser):
 
-        # for --create, add experiment description
+        # for --create, add dataset description
+        parser.add_argument('--desc', type=str, default="", help='Dataset description')
+
         parser.add_argument('--down', action="store_true", default=False, help='Download dataset zipfile')
         parser.add_argument('--unpublish', action="store_true", default=False, help='Unpublish a dataset')
         parser.add_argument('--publish', action="store_true", default=False, help='Publish a public dataset. Makes it available for public download.')
         parser.add_argument('--publish-private', action="store_true", default=False, help='Publish a private dataset. Makes it available for globus download by project collaborators.')
+
+
 
     def down(self, objects, args, out=sys.stdout):
         """Download dataset zipfile, --down
@@ -192,11 +95,11 @@ class DatasetSubcommand(ListObjects):
             out.write("Title: " + obj['title'] + "\n")
             out.write("ID: " + obj['id'] + "\n")
             out.write("Downloading...\n")
-            download_dataset_zipfile(obj['id'], obj['id']+".zip", proj.remote)
+            mcapi.download_dataset_zipfile(obj['id'], obj['id']+".zip", proj.remote)
             out.write("DONE\n\n")
         return
 
-    def create(self, args):
+    def create(self, args, out=sys.stdout):
         """Create new dataset
 
         Using `mc dataset <dataset_name> [--desc <dataset description>]
@@ -208,16 +111,17 @@ class DatasetSubcommand(ListObjects):
             print('example: mc dataset DatasetName --create --desc "dataset description"')
             parser.print_help()
             exit(1)
+
+        resulting_objects = []
         for name in args.expr:
-            if name not in expt_names:
-                dataset = create_dataset(name, args.desc, proj.remote)
-                print('Created dataset:', dataset.id)
-            else:
-                print('experiment: \'' + name + '\' already exists')
+            dataset = mcapi.create_dataset(proj.id, name, description=args.desc, remote=proj.remote)
+            print('Created dataset:', dataset.id)
+            resulting_objects.append(dataset)
+        self.output(resulting_objects, args, out=out)
         return
 
     def delete(self, objects, args, dry_run, out=sys.stdout):
-        """Create new dataset
+        """Delete datasets
 
         Using:
             mc dataset --id <dataset_id> --proj --delete
@@ -232,7 +136,7 @@ class DatasetSubcommand(ListObjects):
 
         for obj in objects:
             try:
-                delete_dataset(proj.id, obj.id, proj.remote)
+                mcapi.delete_dataset(proj.id, obj.id, remote=proj.remote)
             except requests.exceptions.HTTPError as e:
                 try:
                     print(e.response.json()["error"])
@@ -253,7 +157,12 @@ class DatasetSubcommand(ListObjects):
         resulting_objects = []
         for obj in objects:
             try:
-                resulting_objects.append(unpublish_dataset(proj.id, obj.id, proj.remote))
+                # current situation is private datasets get returned after unpublishing
+                # but public datasets return nothing
+                res = mcapi.unpublish_dataset(proj.id, obj.id, proj.remote)
+                if not res:
+                    res = mcapi.get_dataset(proj.id, obj.id, proj.remote)
+                resulting_objects.append(res)
             except requests.exceptions.HTTPError as e:
                 try:
                     print(e.response.json()["error"])
@@ -275,7 +184,7 @@ class DatasetSubcommand(ListObjects):
         resulting_objects = []
         for obj in objects:
             try:
-                resulting_objects.append(publish_dataset(proj.id, obj.id, proj.remote))
+                resulting_objects.append(mcapi.publish_dataset(proj.id, obj.id, proj.remote))
             except requests.exceptions.HTTPError as e:
                 try:
                     print(e.response.json()["error"])
@@ -297,7 +206,7 @@ class DatasetSubcommand(ListObjects):
         resulting_objects = []
         for obj in objects:
             try:
-                resulting_objects.append(publish_private_dataset(proj.id, obj.id, proj.remote))
+                resulting_objects.append(mcapi.publish_private_dataset(proj.id, obj.id, proj.remote))
             except requests.exceptions.HTTPError as e:
                 try:
                     print(e.response.json()["error"])
